requirements.txt
asyncpg==0.22.0
discord.py==2.1.0
cryptography==3.4.6
gspread-asyncio==1.7.1
pycodestyle==2.7.0
python-twitch-client==0.7.1